Qos Service config (#2644)
* Service config

* Allow clients to run without configuring limits

* simpler tests

+++ atlasdb-client/src/main/java/com/palantir/atlasdb/qos/AtlasDbQosClient.java
@@ -24,7 +24,7 @@ import java.util.concurrent.TimeUnit;
 public class AtlasDbQosClient {
     QosService qosService;
 
-    volatile int credits;
+    volatile long credits;
 
     private AtlasDbQosClient(QosService qosService) {
         this.qosService = qosService;
